Create meshtrekker.py
This game is a idea and placeholder as well a start to further implement it into the meshing-around bot setup.

Kelly, the idea is cool, but needs some work. and as well a implementation for the dashboard with maybe a seperated map to track players there movement.  But let's figure this out later hehe

Enjoy the draft!